backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
cryptography
bitarray
sqlalchemy
psycopg[binary]
python-dotenv
httpx
bcrypt
pyjwt
